Now: In Create method MotorcycleController the image, MotorcycleId, Model, Price properties are passed to the method form the view.
Problem:
However The Brand and Category are not sent to The same method from th view.

Solution:
The the first parameter of Create Method has to be of type MotorcycleVM and the created instance of a type has to be of type MotorcycleVM too.
Like this line:      var createModel = new MotorcycleVM();