Problem: HTTPPOST Create() MotorcycleController Brand and Category of the motorcycleViewModel was null after runnig the project.
Solution:
 @Html.DropDownListFor(model => model.Motorcycle.Brand.Name from Create.cshtml
and
ViewBag.BrandId =
        new SelectList(db.Brands, "Name", "Name"); from  HTTPGET Create() are related to each other.
In the mention SelectList the second and third parameter are about what data to be send to @Html.DropDownListFor to show to the view and what data to be recieved from view to HTTPPOST Create(motorcycleViewModel).

Have a look on changes that I made to MotorcyclesController.cs and Create.cshtml to find out more.